taskproject/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==9.2.0
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
